Increase air alarm low pressure thresholds.
If a room's pressure drops it can become unbreathable long before people start taking pressure damage, and so also long before the air alarm starts reporting a danger state. I've increased the values so that the air alarm should show danger if the pressure becomes so low as to be unbreathable.