Basic Debranding (#15)
## About the PR

I believed I've removed all the "player facing" Delta-V branding from
the game. So
- Client name
- Client icons and splash
- Lobby screens

## Why / Balance

The client details were a necessary change for debranding. The current
lobby art has 'Delta-V' across the top, and we'll need to ask the
authors if they want to change it for EE. For now, I moved over the
wizden ones since they're unbranded.

## Media


![logo](https://github.com/Simple-Station/Einstein-Engines/assets/158120145/cbb9b2e4-50e4-4217-ace3-93c33dadeab3)

:cl:
- add: Added a new icon
- tweak: Switched to Wizard's Den's lobby art

---------

Co-authored-by: DEATHB4DEFEAT <77995199+DEATHB4DEFEAT@users.noreply.github.com>