Revert PR #56 through #76 (#104)
# Description

There was a miscommunication, and I bungled the repository during the
last round of merges. The Mirror Bot is not quite finished yet, and some
polishing needs to happen behind the scenes before we can actually start
doing this. Also new policies from the discord are pending regarding how
these PRs need to be handled. I'll need to either reopen or remake the
reverted PRs, since we do actually still need them. But we need PRs
mirrored from Wizards' Den in order to go through the list properly.


![image](https://github.com/Simple-Station/Einstein-Engines/assets/16548818/cb2df7ff-d85b-4653-93a2-d9001e99f175)