Make Debris Actually Dangerous (#353)
# Description

Meant to be a counterpart to this PR:
https://github.com/Simple-Station/Einstein-Engines/pull/342

This makes it so that glass shards and floor tiles will embed in people
when colliding with them, dealing a small amount of damage. This is also
helpful for both making Monstermos with Physics Based Air Throws both
less dangerous, and more interesting. Now if an explosion turns a room
into a whirlwind of scrap metal, people will actually be significantly
concerned about trying not to get hit by debris.

# Changelog

:cl:
- add: Glass/Crystal shards and floor tiles now deal damage and embed in
victims when colliding with them.