Harpy Peacock Tail (#428)
Courtesy of @StillIcarus

# Description

This adds a new Peacock tail marking to Harpies. I'm also taking this
time to update them to the recent DeltaV changes to the birbs.

# Media

![image](https://github.com/Simple-Station/Einstein-Engines/assets/16548818/4f776006-700e-4d92-802c-03e6dd201491)

![image](https://github.com/Simple-Station/Einstein-Engines/assets/16548818/5fd143ab-069f-4d1c-b5e0-ee982a95b4b6)

The backside is semi cursed, but that's something I can't do within the
scope of this PR due to sprite system limitations that Death has planned
improvements for. Just ignore them. :)

Have a birb as consolation.

![faridaiscute](https://github.com/Simple-Station/Einstein-Engines/assets/16548818/09373a57-f32f-48eb-b52b-6d12143f62ff)

:cl: VMSolidus
- add: Peacock Tails have been added for Harpies

---------

Signed-off-by: VMSolidus <evilexecutive@gmail.com>
Co-authored-by: Danger Revolution! <142105406+DangerRevolution@users.noreply.github.com>