Port Randomly Fire Dropped Weapons (#471)
# Description

Ports
https://github.com/Simple-Station/Parkstation-Friendly-Chainsaw/pull/19

I have added a new stat for firearms, the "Reliability" stat, which is a
number between 0 and 1. It's used as a percentage chance for the weapon
to fire itself when violently thrown into anyone. This PR differs from
the original one slightly in that to get it to actually work without
crashing, I set the system to listen to an event that triggers whenever
the gun collides with another entity, not necessarily just the floor.
This is the same event responsible for the clown's cream pie system, or
for glass shards embedding in an entity.

# Changelog

:cl:
- add: NanoTrasen has disabled the unneeded safeties on your guns- Make
sure you're careful with them!
- tweak: All Firearms now have a reliability stat, some are more
reliable than others. The more reliable a weapon is, the less likely it
is to accidentally discharge when yeeted.