Make Spent Casing Ejection Not Suck (#478)
# Description

Part of #467, #460, and #474

This is a small PR that corrects a math error in SharedGunSystem,
causing shell casings to be "Thrown" directly downwards instead of in an
actually cinematic and exciting arc. While I'm at it, I also corrected
the fixture of base shell casings to favor "Bounciness", and decreased
its mass to approximately 100 grams. Finally, I added a sound for when
casings bounce off of walls, which wasn't present before.


https://github.com/Simple-Station/Einstein-Engines/assets/16548818/56bb4ecc-d5eb-4b36-853b-42f05374150d

:cl:
- fix: Spent bullet casings now fly away from a shooter in a cinematic
manner, rather than fall at their feet.

Co-authored-by: Danger Revolution! <142105406+DangerRevolution@users.noreply.github.com>