More Loadout Items Wave 1 (#425)
# Description
The original release of Loadouts to EE did not have much in the way of
actual implementation as far as the item list was concerned. This PR
seeks to get a head start at fixing that by implementing a fairly
sizeable chunk of new items that can be bought with loadout points.

One important addition is the use of "non-clothing-job-related items",
which helps further set this version of Loadouts apart from the /tg/
style of loadouts implemented elsewhere. One of the key strengths of
this loadout system is that it's not limited to just replacing a job's
clothing selections, but can also add items unrelated to clothing. My
given example for this is the new option for Security
characters(provided they have sufficient number of hours as a Security
Officer) can spend roughly half their loadout points to start with a
spare magazine for their handgun.

Secondly, we also now have an opportunity to place non-nanotrasen
corporate items in the game, in such a way that players can tie their
characters to other corporations. Ostensibly this is to build up to the
idea of "Contractor" characters. I hope that this will help broaden
people's horizons.

# On "Syndie" items
I'd like to take this as an opportunity to kind of push people away from
the whole thing of "THATS A SYNDICATE ITEM!!!!" wherever possible. For
the most part this is very much NRP behavior that I'd like to
discourage. With some notable exceptions, this doesn't mean outright
antag items should be obtainable through Loadouts. To that end, items
such as Interdyne branded smokes, Cybersun branded clothing, and other
items with "In-Universe Company Brands" should be moved away from their
identity as strictly antag objects, and instead things characters could
reasonably have bought with their paycheck.

# Technical Changes
I had to make some improvements to the way Loadouts handles species,
namely it wasn't previously possible to have an item require any entry
from a list of species. I updated it so that SpeciesRequirement takes a
list of Species, and will accept any entry from that list. My given
example for this is the entries for Nitrogen air tanks, which can only
be purchased in loadouts by SlimePersons and Vox(who are not currently
playable, but I am accounting for them). Other usage of this is in the
Oni-specific Security items. I do hope to use this as an opportunity to
introduce more species-specific goods in the future.

# Media

Outerwear

![image](https://github.com/Simple-Station/Einstein-Engines/assets/16548818/8d0d9286-7758-45f3-aaff-506dddd35da6)

More job-related selections available to some departments

![image](https://github.com/Simple-Station/Einstein-Engines/assets/16548818/7b214593-7111-42d7-966e-2a340c766f61)

More variety of trinkets and items, including instruments!

![image](https://github.com/Simple-Station/Einstein-Engines/assets/16548818/1b7befaf-c94c-4238-b93b-cbdf14e939dd)

Species Specific Items!

![image](https://github.com/Simple-Station/Einstein-Engines/assets/16548818/e4a2a7e6-085d-4e1b-9ed9-635af8b70e84)

# Changelog

:cl: VMSolidus
- add: The catalog of items available via Loadouts has been greatly
expanded. Have fun customizing your characters!

---------

Signed-off-by: VMSolidus <evilexecutive@gmail.com>
Signed-off-by: DEATHB4DEFEAT <77995199+DEATHB4DEFEAT@users.noreply.github.com>
Co-authored-by: Danger Revolution! <142105406+DangerRevolution@users.noreply.github.com>
Co-authored-by: SimpleStation14 <130339894+SimpleStation14@users.noreply.github.com>
Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>
Co-authored-by: SimpleStation Changelogs <SimpleStation14@users.noreply.github.com>
Co-authored-by: Pspritechologist <81725545+Pspritechologist@users.noreply.github.com>
Co-authored-by: DEATHB4DEFEAT <77995199+DEATHB4DEFEAT@users.noreply.github.com>
Co-authored-by: stellar-novas <stellar_novas@riseup.net>