More Gamemodes (#527)
# Description
Adds traitors and hellshift to the list of gamemodes offered by the
gamemode vote.

Traitors just make sense.

Hellshift is a new version of survival. It starts out roughly as slow as
survival, but it slowly devolves into a torrent of station events as
frequent as one every 10-30 seconds. Good for when there's no other
~~self-antaggers~~ players to spice up the shift, or when there's one
too many security officers to deal with all the threats this brings.

# Technical details
Had to refactor a lot of RampingStationEventScheduler. Firstly, it was
total shitcode, secondly, it had everything hardcoded, thirdly, it had a
bug that made it possible for events to appear with little to no delay
whatsoever. Now its component features four data fields that can be used
to configure the scheduler and its system contains cleaner and more
readable code.

<!--
This is default collapsed, readers click to expand it and see all your
media
The PR media section can get very large at times, so this is a good way
to keep it clean
The title is written using HTML tags
The title must be within the <summary> tags or you won't see it
-->

<details><summary><h1>Media</h1></summary>
<p>


![image](https://github.com/Simple-Station/Einstein-Engines/assets/69920617/e9831b63-0339-4849-981a-bdc58c7beabe)


![image](https://github.com/Simple-Station/Einstein-Engines/assets/69920617/bfa117d3-cb66-456b-9425-1b3bf80cd60f)


</p>
</details>

# Changelog
:cl:
- tweak: Players can now vote on two more gamemodes: Traitors and
Hellshift.

---------

Signed-off-by: Mnemotechnican <69920617+Mnemotechnician@users.noreply.github.com>
Co-authored-by: DEATHB4DEFEAT <77995199+DEATHB4DEFEAT@users.noreply.github.com>