Fix the Thieving Trait Not Working (#552)
<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

# Description

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->
Fixes #529 
A crash course that I need to check when a component is networked. Now
actually replicates pocket vision to the client.

# Changelog

:cl:
- fix: Fixed thieving trait not granting pocket vision