Add Species Restrictions to Loadout Items (#554)
# Description

This is kind of a small PR here, all I'm really doing is adding
species(read: Harpy) restrictions to all items that are invalid for a
Harpy to wear, and then reparenting items that have a skirt but aren't
tagged as such, so that Harpies can wear them. This now makes widespread
actual usage of species restrictions in Loadouts. And also this stops
people complaining that Harpies can buy items that they cannot wear.


![image](https://github.com/user-attachments/assets/d90a3fdc-8b2c-4fa9-938b-bedcfac0dc7b)

Harpies coincidentally now have a blank Shoes tab.


![image](https://github.com/user-attachments/assets/f3585a90-3518-4288-b305-14c1a675c88d)


# Changelog

:cl:
- fix: Harpies can no longer buy Loadout items that are impossible for
them to wear due to having digitigrade legs.