People Can't Talk In Space (#574)
# Description

This fixes a bug that has been present in the game for ages, that people
can talk to each other in the cold vacuum of space without the aid of a
radio. I figured out that this doesn't actually need me to make my
"SpaceAudioSystem" just yet, and can be handled with an extremely easy
check of the voice source's Transform. Simply put, if they're not
touching any grid, their voice can't be heard. Additionally, if a
speaker is standing on a different grid from a potential receiver, their
voice cannot be heard since the two are on different grids. Effectively,
all verbal communication MUST be done through a radio if anyone involved
is in space, or on a different grid. Voice can no longer travel between
grids.

Sound should not travel through the vacuum of space. Use a radio to
communicate, or physically grab them, and hug them to within 3/10th of a
tile. This extremely small distance allotment is there to simulate
characters "Grabbing" each other in space to enable direct speaking
communication.

This also means that in space, you can have a completely private
conversation with someone by hugging them. Nukies would be able to
discuss their attack plans while in the shadow of a station, without
fear of being overheard by crew in defiance of really basic physics.

**I am coming for general audio next. We will restore the rule of
physics and sanity to this game**

# Media


https://github.com/user-attachments/assets/16f107f5-86ab-480b-85df-92b79632e3c1

# Changelog

:cl:
- add: Voice will no longer travel through space. If you are in space,
you must either have a radio to communicate with others, or get within
hugging distance of someone to talk.