Restore HoP As A Functional Job (#576)
# Description

By request from @OldDanceJacket, this PR does two things. First, it
gives the HoP access to Cargo, so that in ODJ's own words, "The Hop can
be like the Warden to Cargo. They are Brothers In Bureaucracy". The
second thing this PR does is add BASIC department PDAs for each
department to the PTECH vendor, enabling the HOP to hire people on as
the basic roles for every department, with additional PDAs for the
teaching roles(Interns).

By design, all specialized department roles(Such as Chemistry) cannot be
granted by the HOP, and must instead be given by their respective heads
of staff.

# Media


![image](https://github.com/user-attachments/assets/601cee43-f23a-4119-9e5a-e7f987fbd6d8)


# Changelog

:cl:
- add: The Head Of Personnel can now hire people onto basic roles for
each department by vending PDAs.