Fix Subfloors For Diggable Diles (#577)
# Description

By request from @OldDanceJacket , this restores the ability for pipes
and wires to be hidden underneath grass and asphalt. I was originally
going to also port digging, but it turns out that SOMEONE already ported
digging, but completely forgot to also fix the subfloor nonsense.

# Media

With ShowSubfloor:

![image](https://github.com/user-attachments/assets/6aa5fee8-a670-4d3a-8d90-d17c964d59a8)

Subfloors hidden:

![image](https://github.com/user-attachments/assets/df3581e3-b8a2-4020-92f5-30af6f0e7c6e)

Dig the grass to expose pipes and wires:

![image](https://github.com/user-attachments/assets/ecf27a56-cb1d-4666-a235-f8ff94d4dfa2)

# Changelog

:cl:
- fix: Pipes and Wires are now correctly hidden by grass and asphalt
once more. Use a shovel to expose pipes hidden by grass.