Fix Talking Across Docked Grids (#593)
# Description

https://github.com/Simple-Station/Einstein-Engines/pull/574 introduced a
bug whereby people who are each on separate grids, but the grids were
attached, could not talk to each other. I have corrected this by making
it so that when the Cross-Grid check is handled, it also checks to see
if the two different grids are JOINTed to each other. Therefore allowing
sound to travel across the connection. This should also work from
Shuttle To Planet, since it's also handled via the same system. This
means that docked shuttles allow sound to travel across to the docked
station(or other shuttle), as well as shuttles docked to a planet's
surface will permit sound to travel to the planet.

# Changelog

:cl:
- fix: Fixed a bug where sound was not traveling over a Shuttle-Docking
connection. Attached grids now permit sound to travel through.

---------

Signed-off-by: VMSolidus <evilexecutive@gmail.com>
Co-authored-by: DEATHB4DEFEAT <77995199+DEATHB4DEFEAT@users.noreply.github.com>