Add Makeup (#614)
<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

# Description

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->

Adds new markings representing makeup:
- Lips
- Blush
- Nail Polish (Left)
- Nail Polish (Right)

Availability depends on the species. Cherry-picked from this
[pull-request](https://github.com/DeltaV-Station/Delta-v/pull/1191) from
Delta-V.

---

<!--
This is default collapsed, readers click to expand it and see all your
media
The PR media section can get very large at times, so this is a good way
to keep it clean
The title is written using HTML tags
The title must be within the <summary> tags or you won't see it
-->

<details><summary><h1>Media</h1></summary>
<p>

Human:


![image](https://github.com/user-attachments/assets/947f0186-297a-4af8-87e4-a14aa08d6100)

![image](https://github.com/user-attachments/assets/fc8b347d-a4fd-49f6-ba8c-2dcd9e5d7ac0)

![image](https://github.com/user-attachments/assets/ae5fc444-b1c2-45bb-969c-d0052c295a9d)


![image](https://github.com/user-attachments/assets/47a1650f-d2a6-4586-a4a1-83398db138b3)

Felinid:


![image](https://github.com/user-attachments/assets/fbd12ed4-e7ac-4898-aa01-bfeff23c2818)

![image](https://github.com/user-attachments/assets/3401464d-ba4d-4da0-b378-b0d45424d792)

![image](https://github.com/user-attachments/assets/f0a08c2f-0cb0-4349-976c-15c59723a702)

Moth:


![image](https://github.com/user-attachments/assets/e0036568-8860-4daf-b548-3f953f9a2ec4)

![image](https://github.com/user-attachments/assets/68620d38-6365-477d-be90-6b92dfaa32ee)
</p>
</details>

---

# Changelog

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl:
- add: Makeup is finally here: lips, blush, and nail polish! Sashay over
to Character Setup in the Markings section, then look at Head/Overlay to
give makeovers to your characters!