Zombie Improvements Take 2 (#652)
# Description

Problem: "Players like to stall zombie rounds"
Solution: Fix the Anti-Stall mechanic.

The set of functions for handling the automatic shuttle call now
correctly only checks "The" station, rather than any grid containing a
StationDataComponent. By using the less permissive
_gameTicker.GetSpawnableStations(); function, the zombie gamerule no
longer includes stations such as Centcomm, Planet Expeditions, some
salvage wrecks, certain space ships, the Syndicate Listening Outpost,
and Nukie World. Additionally, the check for "Healthy" humans no longer
considers anyone infected with the zombie virus to be a "Healthy" human,
ensuring that someone with a 200u jug of Bicaridine can no longer
contribute to round stalling. Finally, dead (player)zombies are now by
default checked for the purpose of calling the evac shuttle. The reason
being for this assumption, is that dead zombies are still players
removed from the round. If a majority of players are removed from the
round(or fucked off to space), the shuttle should be called.

# Changelog

:cl:
- fix: Zombie events have had their Anti-Stalling mechanic improved.
Dead (Player) Zombies, Infected Players, and Initial Infected are all
counted as zombies for the purpose of determine if the shuttle should be
called. Additionally, any player who leaves the station is no longer
counted as a healthy crewman for the automatic shuttle call.