New Trait: Light Step (#658)
# Description

**Light Step** is a 1-point Auditory trait that reduces the volume of
your footsteps by roughly 50%, enabling you to be quieter and
stealthier. Inspired by the SS13 trait of the same name.

## TODO

- [x] Locale strings
- [x] ~~Reduce the distance in which Light Step footsteps can be heard~~
- It turns out that whatever I did reduced the delay between footstep
sounds.
- [x] Discuss the balancing values for the trait

## Technical details

In the current implementation, Light Step reduces the volume of
footsteps by 10 dB, resulting in roughly halving the volume.

The fields in `FootstepVolumeModifierComponent` change the volume in
terms of decibels.

## Media


![image](https://github.com/user-attachments/assets/d1e3efee-6f81-4212-a745-ae08687afcc0)

# Changelog

:cl: Skubman
- add: Add the Light Step trait, a 1-point trait that makes your
footsteps quieter.

---------

Signed-off-by: Angelo Fallaria <ba.fallaria@gmail.com>