Respace Psionics Audio & Localization Files (#700)
# Description

Effectively Part 2 of splitting up the Psionic Refactor. This PR is
partially respacing the non-C# side of Psionics code, such as
Localizations, Audio files, some entities. I'm not respacing all of them
in a single sitting because just finding them all is a complete pain in
the ass.

No media because there's nothing to show.

No changelog because this isn't player facing.